+++ screens/solutions/jogging/Count.tsx
@@ -1,0 +1,51 @@
+
+// import { View } from '../../../components/Themed';
+// import React, { useState, useEffect } from 'react';
+// import { Button, Text } from 'react-native';
+// import * as NativeStack from '@react-navigation/native-stack';
+// import { RootStackParamList } from '../../../types';
+// import { ZStack, HStack, VStack, Spacer } from 'react-native-stacks';
+// import { ModuleBlock } from 'typescript';
+// // export default function Jogging({ navigation }: NativeStack.NativeStackScreenProps<RootStackParamList, 'Jogging'>) {
+// //   return (
+// //     <View style={styles.container}>
+// //       <Button color = "green" title='ホーム画面へ' onPress={() => navigation.goBack()} />
+// //     </View>
+// //   );
+// // }
+// export const Count = ({count}:{count:number}) => {
+
+//   return (
+//     <VStack spacing={10} style={style.stack}>
+//       <Text style={style.time}>{count}</Text>
+//     </VStack>
+//   );
+// };
+
+// const style = StyleSheet.create({
+//   stack:{
+//     width: `80%`,
+//   },
+//   time:{
+//     width: '100%',
+//     textAlign: 'center',
+//     fontSize: 100,
+//     paddingTop: 10,
+//     paddingBottom: 10,
+//     backgroundColor: 'black',
+//     color: 'white',
+//   },
+//   button:{
+//     borderColor: 'black',
+//     borderWidth: 1,
+//     paddingTop: 20,
+//     paddingRight: 40,
+//     paddingBottom: 20,
+//     paddingLeft: 40,
+
+//   },
+//   text:{
+//     fontSize: 30
+//   }
+// });
+// import { StyleSheet, TouchableOpacity } from 'react-native';
